refactor: reorganize components and enhance game logic in Equation Arena
- Moved screen components to a new directory structure for better organization.
- Introduced a new victory sequence handler to manage game state transitions upon enemy defeat.
- Updated reactive blocks to handle game results and trigger appropriate sequences based on game status.
- Added a crafter level configuration system to manage crafting requirements and validations.
- Refactored UI components to improve layout and integrate new display elements for enemies and equations.
