+++ src/lib/minigames/math/EquationArena/components/screens/ResultsScreen.svelte
