+++ src/lib/minigames/math/EquationArena/components/screens/StartScreen.svelte
